+++ R/check_model.R
@@ -9,6 +9,7 @@
 #'
 #' @param x A model object.
 #' @param dot_size,line_size Size of line and dot-geoms.
+#' @param base_size Base font size for plots.
 #' @param panel Logical, if `TRUE`, plots are arranged as panels; else,
 #' single plots for each diagnostic are returned.
 #' @param check Character vector, indicating which checks for should be performed
@@ -191,6 +192,7 @@ check_model.default <- function(x,
                                 dot_alpha = 0.8,
                                 colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                 theme = "see::theme_lucid",
+                                base_size = 10,
                                 detrend = TRUE,
                                 show_dots = NULL,
                                 bandwidth = "nrd",
@@ -261,6 +263,7 @@ check_model.default <- function(x,
   attr(assumptions_data, "panel") <- panel
   attr(assumptions_data, "dot_size") <- dot_size
   attr(assumptions_data, "line_size") <- line_size
+  attr(assumptions_data, "base_size") <- base_size
   attr(assumptions_data, "check") <- check
   attr(assumptions_data, "alpha") <- alpha
   attr(assumptions_data, "dot_alpha") <- dot_alpha
@@ -308,6 +311,7 @@ check_model.stanreg <- function(x,
                                 dot_alpha = 0.8,
                                 colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                 theme = "see::theme_lucid",
+                                base_size = 10,
                                 detrend = TRUE,
                                 show_dots = NULL,
                                 bandwidth = "nrd",
@@ -324,6 +328,7 @@ check_model.stanreg <- function(x,
     dot_alpha = dot_alpha,
     colors = colors,
     theme = theme,
+    base_size = base_size,
     detrend = detrend,
     show_dots = show_dots,
     bandwidth = bandwidth,
@@ -349,6 +354,7 @@ check_model.model_fit <- function(x,
                                   dot_alpha = 0.8,
                                   colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                   theme = "see::theme_lucid",
+                                  base_size = 10,
                                   detrend = TRUE,
                                   show_dots = NULL,
                                   bandwidth = "nrd",
@@ -366,6 +372,7 @@ check_model.model_fit <- function(x,
     dot_alpha = dot_alpha,
     colors = colors,
     theme = theme,
+    base_size = base_size,
     detrend = detrend,
     show_dots = show_dots,
     bandwidth = bandwidth,
@@ -387,6 +394,7 @@ check_model.performance_simres <- function(x,
                                            dot_alpha = 0.8,
                                            colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                            theme = "see::theme_lucid",
+                                           base_size = 10,
                                            detrend = TRUE,
                                            show_dots = NULL,
                                            bandwidth = "nrd",
@@ -404,6 +412,7 @@ check_model.performance_simres <- function(x,
     dot_alpha = dot_alpha,
     colors = colors,
     theme = theme,
+    base_size = base_size,
     detrend = detrend,
     show_dots = show_dots,
     bandwidth = bandwidth,