+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Improved documentation and new vignettes added.
 
+* `check_model()` gets a `base_size` argument, to set the base font size for plots. 
+
 # performance 0.11.0
 
 ## New supported models